[bug:folderstruct done, but id cannot be exported to other file]
