+++ client/src/components/chat/chatDetail.scss
@@ -11,6 +11,12 @@
     border-bottom: 0.5px solid #8080803d;
     color: $orange;
     margin-bottom: 1em;
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    z-index: 1;
+    background: white;
 
     a {
       &:nth-of-type(2) {
@@ -65,7 +71,7 @@
       align-items: flex-start;
       li {
         list-style: none;
-        margin: 1em 0;
+        margin: .5em 0;
 
         div {
           // width: fit-content;
