+++ server/app/routes/router.js
@@ -1,13 +1,14 @@
 const express = require('express');
 const router = express.Router();
-const controllers = require('../controllers/user.controller.js');
+const userControllers = require('../controllers/user.controller.js');
+const chatControllers = require('../controllers/chat.controller.js');
 
 router
-  .post('/api/login', controllers.handleLogin) // Login
-  .post('/api/register', controllers.handleRegister) // Register
-  .get('/api/users/:id', controllers.handleUsers) // Users
-  .get('/api/user/:id', controllers.handleUser) // User
-  .get('/api/chatsItems/:id', controllers.handleChats) // Chats
-  .get('/api/createChat/:id', controllers.handleCreateChat); // Create chat env
+  .post('/api/login', userControllers.handleLogin) // Login
+  .post('/api/register', userControllers.handleRegister) // Register
+  .get('/api/users/:id', userControllers.handleUsers) // Users
+  .get('/api/user/:id', userControllers.handleUser) // User
+  .get('/api/chatsItems/:id', chatControllers.handleChats) // Chats
+  .get('/api/createChat/:id', chatControllers.handleCreateChat); // Create chat env
 
 module.exports = router;
